requirements/prod.txt
-r base.txt

gunicorn==19.7.1
raven==6.1.0
boto==2.47.0
django-storages==1.5.2
datadog==0.16.0
django-otto-admin==0.1.2
django-pipeline==1.6.13
pyOpenSSL==17.2.0
